django/requirements.txt
django>=5.1.0
djangorestframework>=3.15.0
django-cors-headers>=4.6.0
python-dotenv>=1.0.0
